Install s.el and f.el manually instead of using package-vc-install
This should reduce byte-compiler noise: package-vc-install clones the
entire repository and attempts to compile EVERYTHING in it, which can
generate some not-so-useful error messages.

This commit updates the code to fetch just the interest elisp files and
put them somewhere in the load-path. It also installs dash.el using
package-install since it is not pulled as a dependency anymore.